Time: 93 ms (71.62%), Space: 99.1 MB (16.98%) - LeetHub

+++ 3332-minimum-operations-to-exceed-threshold-value-ii/3332-minimum-operations-to-exceed-threshold-value-ii.cpp
@@ -9,10 +9,10 @@ public:
 
         int cnt = 0;
         while (p.size() > 1 && p.top() < k) { // Ensure at least 2 elements exist
-            int x = p.top(); p.pop();
-            int y = p.top(); p.pop();
+          long long   int x = p.top(); p.pop();
+          long long   int y = p.top(); p.pop();
             
-            long long m = ((x * 2LL)  + y) ; // Avoid overflow
+            long long m = ((x * 2)  + y) ; // Avoid overflow
             
             p.push(m);
             cnt++;
